+++ app/database/migrations/2015_03_21_102919_create_favourites_table.php
@@ -16,7 +16,7 @@ class CreateFavouritesTable extends Migration {
 		{
             $table->increments('id');
             $table->integer('user_id')->unsigned();
-            $table->integer('celebrity_id')->unsigned();
+            $table->bigInteger('celebrity_id')->unsigned();
 
             $table->foreign('user_id')->references('id')->on('users')->onUpdate('cascade')->onDelete('cascade');
             $table->foreign('celebrity_id')->references('id')->on('celebrities')->onUpdate('cascade')->onDelete('cascade');
